Messeji HTTP Client

+++ suripu-core/src/main/java/com/hello/suripu/core/db/mappers/sleep_sounds/DurationMapper.java
@@ -16,6 +16,6 @@ public class DurationMapper implements ResultSetMapper<Duration> {
         return Duration.create(
                 r.getLong("id"),
                 r.getString("name"),
-                r.getLong("duration_seconds"));
+                r.getInt("duration_seconds"));
     }
 }
